refactor(teamSelection): extract TeamSelectionScopePanel view component
Move the ~210-line simulation-scope card (home region/country selects,
simulated-regions and simulated-competitions checkboxes) into a
presentational TeamSelectionScopePanel. State and the toggle handlers
stay in the parent and pass in as props; the panel calls useTranslation
internally. Pure move, no behavior change.

TeamSelection.tsx: 1089 -> 825 lines.
